Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.html because it does not appear in the frame-src directive of the Content Security Policy.
